feat(metadata): adjust iginx start up

+++ core/src/main/java/cn/edu/tsinghua/iginx/metadata/storage/zk/ZooKeeperMetaStorage.java
@@ -25,12 +25,14 @@ import static cn.edu.tsinghua.iginx.metadata.utils.IdUtils.generateId;
 import static cn.edu.tsinghua.iginx.metadata.utils.ReshardStatus.*;
 
 import cn.edu.tsinghua.iginx.conf.ConfigDescriptor;
+import cn.edu.tsinghua.iginx.exception.SessionException;
 import cn.edu.tsinghua.iginx.metadata.cache.IMetaCache;
 import cn.edu.tsinghua.iginx.metadata.entity.*;
 import cn.edu.tsinghua.iginx.metadata.exception.MetaStorageException;
 import cn.edu.tsinghua.iginx.metadata.hook.*;
 import cn.edu.tsinghua.iginx.metadata.storage.IMetaStorage;
 import cn.edu.tsinghua.iginx.metadata.utils.ReshardStatus;
+import cn.edu.tsinghua.iginx.session.Session;
 import cn.edu.tsinghua.iginx.transform.pojo.TriggerDescriptor;
 import cn.edu.tsinghua.iginx.utils.JsonUtils;
 import cn.edu.tsinghua.iginx.utils.Pair;
@@ -45,6 +47,7 @@ import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
 import org.apache.curator.framework.recipes.cache.TreeCache;
 import org.apache.curator.framework.recipes.cache.TreeCacheListener;
+import org.apache.curator.framework.recipes.locks.InterProcessMultiLock;
 import org.apache.curator.framework.recipes.locks.InterProcessMutex;
 import org.apache.curator.retry.RetryForever;
 import org.apache.zookeeper.CreateMode;
@@ -56,9 +59,19 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ZooKeeperMetaStorage.class);
 
+  private static final String NODE_PREFIX = "node";
+
   private static final String IGINX_NODE = "/iginx/node";
 
   private static final String STORAGE_ENGINE_NODE = "/storage/node";
+
+  private static final String IGINX_CONNECTION_NODE = "/iginx-connection/node";
+
+  private static final String IGINX_CONNECTION_LOCK_NODE = "/lock/iginx-connection";
+
+  private static final String STORAGE_CONNECTION_NODE = "/storage-connection/node";
+
+  private static final String STORAGE_CONNECTION_LOCK_NODE = "/lock/storage-connection";
 
   private static final String STORAGE_UNIT_NODE = "/unit/unit";
 
@@ -298,13 +311,13 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
         mutex.release();
       } catch (Exception e) {
         throw new MetaStorageException(
-            "get error when release interprocess lock for " + SCHEMA_MAPPING_LOCK_NODE, e);
+            "get error when release interprocess lock for " + IGINX_LOCK_NODE, e);
       }
     }
   }
 
   @Override
-  public long registerIginx(IginxMeta iginx) throws MetaStorageException {
+  public IginxMeta registerIginx(IginxMeta iginx) throws MetaStorageException {
     InterProcessMutex mutex = new InterProcessMutex(client, IGINX_LOCK_NODE);
     try {
       mutex.acquire();
@@ -318,7 +331,7 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
       IginxMeta iginxMeta =
           new IginxMeta(id, iginx.getIp(), iginx.getPort(), iginx.getExtraParams());
       this.client.setData().forPath(nodeName, JsonUtils.toJson(iginxMeta));
-      return id;
+      return iginxMeta;
     } catch (Exception e) {
       throw new MetaStorageException("get error when load iginx", e);
     } finally {
@@ -346,12 +359,10 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
               iginxMeta = JsonUtils.fromJson(data, IginxMeta.class);
               if (iginxMeta != null) {
                 LOGGER.info(
-                    "new iginx comes to cluster: id = "
-                        + iginxMeta.getId()
-                        + " ,ip = "
-                        + iginxMeta.getIp()
-                        + " , port = "
-                        + iginxMeta.getPort());
+                    "new iginx comes to cluster: id = {} ,ip = {} , port = {}",
+                    iginxMeta.getId(),
+                    iginxMeta.getIp(),
+                    iginxMeta.getPort());
                 iginxChangeHook.onChange(iginxMeta.getId(), iginxMeta);
               } else {
                 LOGGER.error("resolve iginx meta from zookeeper error");
@@ -370,12 +381,10 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
               iginxMeta = JsonUtils.fromJson(data, IginxMeta.class);
               if (iginxMeta != null) {
                 LOGGER.info(
-                    "iginx leave from cluster: id = "
-                        + iginxMeta.getId()
-                        + " ,ip = "
-                        + iginxMeta.getIp()
-                        + " , port = "
-                        + iginxMeta.getPort());
+                    "iginx leave from cluster: id = {} ,ip = {} , port = {}",
+                    iginxMeta.getId(),
+                    iginxMeta.getIp(),
+                    iginxMeta.getPort());
                 iginxChangeHook.onChange(iginxMeta.getId(), null);
               } else {
                 LOGGER.error("resolve iginx meta from zookeeper error");
@@ -428,8 +437,7 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
         byte[] data = this.client.getData().forPath(STORAGE_ENGINE_NODE_PREFIX + "/" + childName);
         StorageEngineMeta storageEngineMeta = JsonUtils.fromJson(data, StorageEngineMeta.class);
         if (storageEngineMeta == null) {
-          LOGGER.error(
-              "resolve data from " + STORAGE_ENGINE_NODE_PREFIX + "/" + childName + " error");
+          LOGGER.error("resolve data from " + STORAGE_ENGINE_NODE_PREFIX + "/{} error", childName);
           continue;
         }
         storageEngineMetaMap.putIfAbsent(storageEngineMeta.getId(), storageEngineMeta);
@@ -452,10 +460,15 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
   }
 
   @Override
-  public long addStorageEngine(StorageEngineMeta storageEngine) throws MetaStorageException {
-    InterProcessMutex mutex = new InterProcessMutex(this.client, STORAGE_ENGINE_LOCK_NODE);
+  public long addStorageEngine(long iginxId, StorageEngineMeta storageEngine)
+      throws MetaStorageException {
+    InterProcessMutex storageMutex = new InterProcessMutex(this.client, STORAGE_ENGINE_LOCK_NODE);
+    InterProcessMutex connectionMutex =
+        new InterProcessMutex(this.client, STORAGE_CONNECTION_LOCK_NODE);
+    InterProcessMultiLock multiLock =
+        new InterProcessMultiLock(Arrays.asList(storageMutex, connectionMutex));
     try {
-      mutex.acquire();
+      multiLock.acquire();
       String nodeName =
           this.client
               .create()
@@ -465,36 +478,84 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
       long id = Long.parseLong(nodeName.substring(STORAGE_ENGINE_NODE.length()));
       storageEngine.setId(id);
       this.client.setData().forPath(nodeName, JsonUtils.toJson(storageEngine));
+      // 记录连接关系
+      String connectionPath = generateId(STORAGE_CONNECTION_NODE, iginxId);
+      long[] ids;
+      if (this.client.checkExists().forPath(connectionPath) == null) {
+        this.client
+            .create()
+            .creatingParentsIfNeeded()
+            .withMode(CreateMode.EPHEMERAL)
+            .forPath(connectionPath);
+        ids = new long[1];
+        ids[0] = id;
+      } else {
+        byte[] data = client.getData().forPath(connectionPath);
+        long[] oldIds = JsonUtils.fromJson(data, long[].class);
+        ids = new long[oldIds.length + 1];
+        System.arraycopy(oldIds, 0, ids, 0, oldIds.length);
+        ids[oldIds.length] = id;
+      }
+      this.client.setData().forPath(nodeName, JsonUtils.toJson(ids));
       return id;
     } catch (Exception e) {
       throw new MetaStorageException("get error when add storage engine", e);
     } finally {
       try {
-        mutex.release();
+        multiLock.release();
       } catch (Exception e) {
         throw new MetaStorageException(
-            "get error when release interprocess lock for " + SCHEMA_MAPPING_LOCK_NODE, e);
+            String.format(
+                "get error when release interprocess lock for %s and %s",
+                STORAGE_ENGINE_LOCK_NODE, STORAGE_CONNECTION_LOCK_NODE),
+            e);
       }
     }
   }
 
   @Override
-  public void removeDummyStorageEngine(long storageEngineId) throws MetaStorageException {
-    InterProcessMutex mutex = new InterProcessMutex(this.client, STORAGE_ENGINE_LOCK_NODE);
+  public void removeDummyStorageEngine(long iginxId, long storageEngineId, boolean forAllIginx)
+      throws MetaStorageException {
+    InterProcessMutex storageMutex = new InterProcessMutex(this.client, STORAGE_ENGINE_LOCK_NODE);
+    InterProcessMutex connectionMutex =
+        new InterProcessMutex(this.client, STORAGE_CONNECTION_LOCK_NODE);
+    InterProcessMultiLock multiLock =
+        new InterProcessMultiLock(Arrays.asList(storageMutex, connectionMutex));
     try {
-      mutex.acquire();
+      multiLock.acquire();
       String nodeName = generateId(STORAGE_ENGINE_NODE, storageEngineId);
-      if (this.client.checkExists().forPath(nodeName) != null) {
+      if (forAllIginx && this.client.checkExists().forPath(nodeName) != null) {
         this.client.delete().forPath(nodeName);
+      }
+      // 删除连接状态
+      String connectionPath = generateId(STORAGE_CONNECTION_NODE, iginxId);
+      if (this.client.checkExists().forPath(connectionPath) != null) {
+        byte[] data = client.getData().forPath(connectionPath);
+        long[] ids = JsonUtils.fromJson(data, long[].class);
+        int index = Arrays.binarySearch(ids, storageEngineId);
+        if (index >= 0) {
+          if (ids.length > 1) {
+            long[] newIds = new long[ids.length - 1];
+            System.arraycopy(ids, 0, newIds, 0, index);
+            System.arraycopy(ids, index + 1, newIds, index, ids.length - index - 1);
+            this.client.setData().forPath(nodeName, JsonUtils.toJson(newIds));
+          } else {
+            // iginx 将没有连接的存储节点
+            this.client.delete().forPath(connectionPath);
+          }
+        }
       }
     } catch (Exception e) {
       throw new MetaStorageException("get error when removing dummy storage engine", e);
     } finally {
       try {
-        mutex.release();
+        multiLock.release();
       } catch (Exception e) {
         throw new MetaStorageException(
-            "get error when release interprocess lock for " + STORAGE_ENGINE_LOCK_NODE, e);
+            String.format(
+                "get error when release interprocess lock for %s and %s",
+                STORAGE_ENGINE_LOCK_NODE, STORAGE_CONNECTION_LOCK_NODE),
+            e);
       }
     }
   }
@@ -520,12 +581,10 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
               storageEngineMeta = JsonUtils.fromJson(data, StorageEngineMeta.class);
               if (storageEngineMeta != null) {
                 LOGGER.info(
-                    "new storage engine comes to cluster: id = "
-                        + storageEngineMeta.getId()
-                        + " ,ip = "
-                        + storageEngineMeta.getIp()
-                        + " , port = "
-                        + storageEngineMeta.getPort());
+                    "new storage engine comes to cluster: id = {} ,ip = {} , port = {}",
+                    storageEngineMeta.getId(),
+                    storageEngineMeta.getIp(),
+                    storageEngineMeta.getPort());
                 storageChangeHook.onChange(storageEngineMeta.getId(), storageEngineMeta);
               } else {
                 LOGGER.error("resolve storage engine from zookeeper error");
@@ -544,12 +603,10 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
               storageEngineMeta = JsonUtils.fromJson(data, StorageEngineMeta.class);
               if (storageEngineMeta != null) {
                 LOGGER.info(
-                    "storage engine leave from cluster: id = "
-                        + storageEngineMeta.getId()
-                        + " ,ip = "
-                        + storageEngineMeta.getIp()
-                        + " , port = "
-                        + storageEngineMeta.getPort());
+                    "storage engine leave from cluster: id = {} ,ip = {} , port = {}",
+                    storageEngineMeta.getId(),
+                    storageEngineMeta.getIp(),
+                    storageEngineMeta.getPort());
                 storageChangeHook.onChange(storageEngineMeta.getId(), null);
               } else {
                 LOGGER.error("resolve storage engine from zookeeper error");
@@ -566,6 +623,148 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
   @Override
   public void registerStorageChangeHook(StorageChangeHook hook) {
     this.storageChangeHook = hook;
+  }
+
+  @Override
+  public Map<Long, Session> connectOtherIginx(IginxMeta iginx) throws MetaStorageException {
+    InterProcessMutex iginxMutex = new InterProcessMutex(client, IGINX_LOCK_NODE);
+    InterProcessMutex connectionMutex = new InterProcessMutex(client, IGINX_CONNECTION_LOCK_NODE);
+    InterProcessMultiLock multiLock =
+        new InterProcessMultiLock(Arrays.asList(iginxMutex, connectionMutex));
+    try {
+      multiLock.acquire();
+      Map<Long, Session> sessionMap = new HashMap<>();
+      List<Long> connectedIds = new ArrayList<>();
+      List<String> children = client.getChildren().forPath(IGINX_NODE_PREFIX);
+      for (String childName : children) {
+        byte[] data = client.getData().forPath(IGINX_NODE_PREFIX + "/" + childName);
+        IginxMeta iginxMeta = JsonUtils.fromJson(data, IginxMeta.class);
+        if (iginxMeta == null) {
+          LOGGER.error("resolve data from {}/{} error", IGINX_NODE_PREFIX, childName);
+          continue;
+        }
+        if (iginx.getIp().equals(iginxMeta.getIp()) && iginx.getPort() == iginxMeta.getPort()) {
+          continue;
+        }
+        // 尝试与集群中其他iginx建立连接
+        Session session = new Session(iginxMeta.iginxMetaInfo());
+        try {
+          session.openSession();
+        } catch (SessionException e) {
+          LOGGER.info(
+              "open session of iginx(id = {} ,ip = {} , port = {}) failed, because: ",
+              iginxMeta.getId(),
+              iginxMeta.getIp(),
+              iginxMeta.getPort(),
+              e);
+          continue;
+        }
+        LOGGER.info(
+            "connect to iginx(id = {} ,ip = {} , port = {})",
+            iginxMeta.getId(),
+            iginxMeta.getIp(),
+            iginxMeta.getPort());
+        sessionMap.putIfAbsent(iginxMeta.getId(), session);
+        connectedIds.add(iginxMeta.getId());
+      }
+      String nodeName =
+          this.client
+              .create()
+              .creatingParentsIfNeeded()
+              .withMode(CreateMode.EPHEMERAL)
+              .forPath(generateId(IGINX_CONNECTION_NODE, iginx.getId()));
+      long[] ids = connectedIds.stream().mapToLong(Long::longValue).toArray();
+      this.client.setData().forPath(nodeName, JsonUtils.toJson(ids));
+      return sessionMap;
+    } catch (Exception e) {
+      throw new MetaStorageException("get error when connect to other iginx", e);
+    } finally {
+      try {
+        multiLock.release();
+      } catch (Exception e) {
+        throw new MetaStorageException(
+            String.format(
+                "get error when release interprocess lock for %s and %s",
+                IGINX_LOCK_NODE, IGINX_CONNECTION_LOCK_NODE),
+            e);
+      }
+    }
+  }
+
+  @Override
+  public Map<Long, List<Long>> updateClusterIginxConnections() throws MetaStorageException {
+    InterProcessMutex mutex = new InterProcessMutex(client, IGINX_CONNECTION_LOCK_NODE);
+    try {
+      mutex.acquire();
+      Map<Long, List<Long>> connectionMap = new HashMap<>();
+      if (this.client.checkExists().forPath(IGINX_CONNECTION_NODE_PREFIX) == null) {
+        // 当前还没有数据，创建父节点，然后不需要解析数据
+        client.create().withMode(CreateMode.EPHEMERAL).forPath(IGINX_CONNECTION_NODE_PREFIX);
+      } else {
+        List<String> children = client.getChildren().forPath(IGINX_CONNECTION_NODE_PREFIX);
+        for (String childName : children) {
+          byte[] data = client.getData().forPath(IGINX_CONNECTION_NODE_PREFIX + "/" + childName);
+          long[] toIds = JsonUtils.fromJson(data, long[].class);
+          if (toIds == null) {
+            LOGGER.error("resolve data from {}/{} error", IGINX_CONNECTION_NODE_PREFIX, childName);
+            continue;
+          }
+          long fromId = Long.parseLong(childName.substring(NODE_PREFIX.length()));
+          for (long toId : toIds) {
+            connectionMap.computeIfAbsent(fromId, k -> new ArrayList<>()).add(toId);
+          }
+        }
+      }
+      return connectionMap;
+    } catch (Exception e) {
+      throw new MetaStorageException("get error when update cluster iginx connections", e);
+    } finally {
+      try {
+        mutex.release();
+      } catch (Exception e) {
+        throw new MetaStorageException(
+            "get error when release interprocess lock for " + IGINX_CONNECTION_LOCK_NODE, e);
+      }
+    }
+  }
+
+  @Override
+  public Map<Long, List<Long>> updateClusterStorageConnections() throws MetaStorageException {
+    InterProcessMutex mutex = new InterProcessMutex(client, STORAGE_CONNECTION_LOCK_NODE);
+    try {
+      mutex.acquire();
+      Map<Long, List<Long>> connectionMap = new HashMap<>();
+
+      if (this.client.checkExists().forPath(STORAGE_CONNECTION_NODE_PREFIX) == null) {
+        // 当前还没有数据，创建父节点，然后不需要解析数据
+        client.create().withMode(CreateMode.EPHEMERAL).forPath(STORAGE_CONNECTION_NODE_PREFIX);
+      } else {
+        List<String> children = client.getChildren().forPath(STORAGE_CONNECTION_NODE_PREFIX);
+        for (String childName : children) {
+          byte[] data = client.getData().forPath(STORAGE_CONNECTION_NODE_PREFIX + "/" + childName);
+          long[] storageIds = JsonUtils.fromJson(data, long[].class);
+          if (storageIds == null) {
+            LOGGER.error(
+                "resolve data from {}/{} error", STORAGE_CONNECTION_NODE_PREFIX, childName);
+            continue;
+          }
+          long iginxId = Long.parseLong(childName.substring(NODE_PREFIX.length()));
+          for (long storageId : storageIds) {
+            connectionMap.computeIfAbsent(iginxId, k -> new ArrayList<>()).add(storageId);
+          }
+        }
+      }
+      return connectionMap;
+    } catch (Exception e) {
+      throw new MetaStorageException("get error when update cluster storage connections", e);
+    } finally {
+      try {
+        mutex.release();
+      } catch (Exception e) {
+        throw new MetaStorageException(
+            "get error when release interprocess lock for " + STORAGE_CONNECTION_LOCK_NODE, e);
+      }
+    }
   }
 
   @Override
@@ -588,9 +787,8 @@ public class ZooKeeperMetaStorage implements IMetaStorage {
                 storageUnitMetaMap.get(storageUnitMeta.getMasterId());
             if (masterStorageUnitMeta == null) { // 子节点先于主节点加入系统中，不应该发生，报错
               LOGGER.error(
-                  "unexpected storage unit "
-                      + new String(data)
-                      + ", because it does not has a master storage unit");
+                  "unexpected storage unit {}, because it does not has a master storage unit",
+                  new String(data));
             } else {
               masterStorageUnitMeta.addReplica(storageUnitMeta);
             }
